Tags: Viz: Fix issues with hovering to highlight tag in viz
Problem was that when you hovered over a tag with subtags (e.g. `sentiment`), it couldn't find any nodes with that ID in the viz because the viz splits them out into subtags. No idea why this made the app hang, but nor should it have worked. This commit refactors things a bunch to pass subtags and subtag IDs through properly. It's pretty hacky, however, since subtags are a total hack right now.