Tooltips: Going gentle with MutationObserver
Initializing tooltips with MutationObserver too early was a) slow, and b) causing Firefox to hang from gajillion events firing as the page loads.

Also now using the pre-built minified version of `darsain-tooltips`, which avoids  massive ugly (and slow?) `eval`ed functions, which are only present for the purpose of supporting source maps or something.